requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
sqlmodel
langchain
langchain-community
langchain-groq
chromadb
sentence-transformers
groq
langdetect
SpeechRecognition
gTTS
streamlit
streamlit-webrtc
plotly
requests
pydub
av
pypdf
langchain-chroma
